Split punktf into two crates (library and cli)
