+++ crates/punktf-lib/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "punktf-lib"
+version = "1.0.0-alpha.1"
+authors = ["Michael Lohr <michael@lohr.dev>", "Shemnei"]
+edition = "2018"
+description = "Library for punktf, a cross-platform multi-target dotfiles manager"
+documentation = "https://github.com/Shemnei/punktf/tree/master/crates/punktf-lib"
+homepage = "https://github.com/Shemnei/punktf/tree/master/crates/punktf-lib"
+repository = "https://github.com/Shemnei/punktf/tree/master/crates/punktf-lib"
+license= "MIT"
+keywords = ["dotfiles", "cli", "dotfile", "dotfiles-manager", "templating"]
+
+[lib]
+name = "punktf_lib"
+
+[dependencies]
+color-eyre = "0.5.11"
+env_logger = "0.9.0"
+log = "0.4.14"
+serde = { version = "1.0.126", features = ["derive"] }
+serde_json = "1.0.64"
+serde_yaml = "0.8.17"
+thiserror = "1.0.26"
+walkdir = "2.3.2"
+unicode-width = "0.1.8"
+cfg-if = "1.0.0"
+
+[dev-dependencies]
+pretty_assertions = "0.7.2"
