+++ crates/punktf-lib/src/template/diagnostic.rs
